4/17 work
- fixed the bug from yesterday
- /response loads now but submitting has an internal server error that i'm still figuring out
