+++ app/templates/response.html
@@ -7,7 +7,7 @@
 <body>
 <h1>Email Templates</h1>
 
-<form action= "/app.py" method="post">
+<form action= "{{url_for('index.response')}}" method="post">
 <!-- This is code for each individual response from feid to the customer -->
 <fieldset>  <!-- When it is waiting to be accepted, he can accept or deny it-->
             <legend>Accepted Email Template</legend>
